requirements.txt
numpy>=1.24,<2.0
pandas>=2.0,<3.0
scipy>=1.11
scikit-learn>=1.3
xgboost-cpu>=2.0   # CPU-only wheel; same `import xgboost` namespace, no ~1.5 GB nvidia/CUDA deps
pyarrow>=13.0
pyyaml>=6.0
matplotlib>=3.7
streamlit>=1.30
mlflow==2.16.0
fastapi>=0.110
uvicorn>=0.27
prometheus-client>=0.20
python-dotenv>=1.0
pytest>=8.0
